return evolved nonces
