+++ crates/amaru/src/bin/amaru/cmd/import_nonces.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use amaru_consensus::consensus::store::{ChainStore, Nonces};
+use amaru_consensus::{consensus::store::ChainStore, Nonces};
 use amaru_kernel::{network::NetworkName, Hash, Header, Nonce, Point};
 use amaru_stores::rocksdb::consensus::RocksDBStore;
 use clap::Parser;
@@ -78,7 +78,7 @@ pub async fn run(args: Args) -> Result<(), Box<dyn std::error::Error>> {
         tail: args.tail,
     };
 
-    db.put_nonces(&header_hash, nonces)?;
+    db.put_nonces(&header_hash, &nonces)?;
 
     Ok(())
 }
