+++ crates/ouroboros-traits/src/praos/mod.rs
@@ -45,7 +45,46 @@ Summarizing:
 */
 
 use crate::is_header::IsHeader;
-use amaru_kernel::{Hash, Nonce};
+use amaru_kernel::{cbor, Epoch, Hash, Nonce};
+
+#[derive(Debug, PartialEq, Clone)]
+pub struct Nonces {
+    pub active: Nonce,
+    pub evolving: Nonce,
+    pub candidate: Nonce,
+    pub tail: Hash<32>,
+    pub epoch: Epoch,
+}
+
+impl<C> cbor::encode::Encode<C> for Nonces {
+    fn encode<W: cbor::encode::Write>(
+        &self,
+        e: &mut cbor::Encoder<W>,
+        ctx: &mut C,
+    ) -> Result<(), cbor::encode::Error<W::Error>> {
+        e.begin_array()?;
+        e.encode_with(self.active, ctx)?;
+        e.encode_with(self.evolving, ctx)?;
+        e.encode_with(self.candidate, ctx)?;
+        e.encode_with(self.tail, ctx)?;
+        e.encode_with(self.epoch, ctx)?;
+        e.end()?;
+        Ok(())
+    }
+}
+
+impl<'b, C> cbor::decode::Decode<'b, C> for Nonces {
+    fn decode(d: &mut cbor::Decoder<'b>, ctx: &mut C) -> Result<Self, cbor::decode::Error> {
+        d.array()?;
+        Ok(Nonces {
+            active: d.decode_with(ctx)?,
+            evolving: d.decode_with(ctx)?,
+            candidate: d.decode_with(ctx)?,
+            tail: d.decode_with(ctx)?,
+            epoch: d.decode_with(ctx)?,
+        })
+    }
+}
 
 pub trait Praos<H: IsHeader>: Send + Sync {
     type Error;
@@ -63,5 +102,5 @@ pub trait Praos<H: IsHeader>: Send + Sync {
     ///
     /// Once the stability window has been reached, the candidate is fixed for the epoch and will
     /// be used once crossing the epoch boundary to produce the next epoch nonce.
-    fn evolve_nonce(&mut self, header: &H) -> Result<(), Self::Error>;
+    fn evolve_nonce(&mut self, header: &H) -> Result<Nonces, Self::Error>;
 }
